+++ content/math/sumFloorArithSeries.h
@@ -7,12 +7,12 @@
 //(p, q, n are natural numbers)
 //[x] = floor(x)
 
-ll cnt(ll p, ll q, ll n) {
-	ll t = __gcd(p, q);
+long long cnt(long long p, long long q, long long n) {
+	long long t = __gcd(p, q);
 	p = p / t;
 	q = q / t;
-	ll s = 0;
-	ll z = 1;
+	long long s = 0;
+	long long z = 1;
 	while ((q > 0) && (n > 0)) {
 		//(point A)
 		t = p / q;
